requirements.txt
numpy==1.25.2
pygame-ce==2.3.1
pygame-gui==0.6.9
python-i18n==0.3.9
scipy==1.11.2
